--- client/src/cl_http.h
@@ -1,41 +1,0 @@
-// Emacs style mode select   -*- C++ -*-
-//-----------------------------------------------------------------------------
-//
-// $Id$
-//
-// Copyright (C) 2006-2020 by The Odamex Team.
-//
-// This program is free software; you can redistribute it and/or
-// modify it under the terms of the GNU General Public License
-// as published by the Free Software Foundation; either version 2
-// of the License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// DESCRIPTION:
-//	HTTP Downloading.
-//
-//-----------------------------------------------------------------------------
-
-#ifndef __CL_HTTP_H__
-#define __CL_HTTP_H__
-
-#include <string>
-
-namespace http
-{
-
-void Init();
-void Shutdown();
-bool IsDownloading();
-void StartDownload(const std::string& website, const std::string& filename,
-                   const std::string& hash);
-void Tick();
-std::string Progress();
-
-} // namespace http
-
-#endif
