+++ server/src/sv_main.cpp
@@ -4532,161 +4532,6 @@ void SV_WinCheck (void)
 EXTERN_CVAR (sv_waddownloadcap)
 EXTERN_CVAR (sv_download_test)
 
-void SV_OptimizedDownloadTest()
-{
-	// nobody around?
-	if (players.empty())
-		return;
-
-	FILE* file = NULL;
-	std::string lastDownloadName;
-
-	// wad downloading
-	for (Players::iterator it = players.begin(); it != players.end(); ++it)
-	{
-		if (it->playerstate != PST_DOWNLOAD)
-			continue;
-
-		client_t *cl = &(it->client);
-
-		if (!cl->download.name.length())
-			continue;
-
-		static char buff[1024];
-		// Smaller buffer for slower clients
-		int chunk_size = (sizeof(buff) > (unsigned)cl->rate * 1000 / TICRATE) ? cl->rate * 1000 / TICRATE : sizeof(buff);
-
-		// maximum rate client can download at (in bytes per second)
-		int download_rate = (sv_waddownloadcap > cl->rate) ? cl->rate * 1000 : sv_waddownloadcap * 1000;
-
-		//Try to open/close the download file as little as possible.
-		//Too many fopen/fclose calls is inefficient and can choke out the game loop causing the server to be choppy with multiple clients downloading
-		if (lastDownloadName != cl->download.name)
-		{
-			if (file)
-				fclose(file);
-			file = fopen(cl->download.name.c_str(), "rb");
-			lastDownloadName = cl->download.name;
-		}
-
-		if (!file)
-			return;
-
-		do
-		{
-			// [SL] 2011-08-09 - Always send the data in netbuf and reliablebuf prior
-			// to writing a wadchunk to netbuf to keep packet sizes below the MTU.
-			// This prevents packets from getting dropped due to size on some networks.
-			if (cl->netbuf.size() + cl->reliablebuf.size())
-				SV_SendPacket(*it);
-
-			if (!cl->download.next_offset)
-			{
-				MSG_WriteMarker(&cl->netbuf, svc_wadinfo);
-				MSG_WriteLong(&cl->netbuf, M_FileLength(file));
-			}
-
-			fseek(file, cl->download.next_offset, SEEK_SET);
-			size_t read = fread(buff, 1, chunk_size, file);
-
-			if (!read)
-				break;
-
-			MSG_WriteMarker(&cl->netbuf, svc_wadchunk);
-			MSG_WriteLong(&cl->netbuf, cl->download.next_offset);
-			MSG_WriteShort(&cl->netbuf, read);
-			MSG_WriteChunk(&cl->netbuf, buff, read);
-
-			// Make double-sure the wadchunk is sent in its own packet
-			if (cl->netbuf.size() + cl->reliablebuf.size())
-				SV_SendPacket(*it);
-
-			cl->download.next_offset += read;
-		} while (
-			(double)(cl->reliable_bps + cl->unreliable_bps) * TICRATE
-			/ (double)(gametic % TICRATE)	// bps already used
-			+ (double)chunk_size / TICRATE 	// bps this chunk will use
-			< (double)download_rate);
-	}
-
-	if (file)
-		fclose(file);
-}
-
-void SV_DownloadOriginal()
-{
-	// nobody around?
-	if (players.empty())
-		return;
-
-	// wad downloading
-	for (Players::iterator it = players.begin(); it != players.end(); ++it)
-	{
-		if (it->playerstate != PST_DOWNLOAD)
-			continue;
-
-		client_t *cl = &(it->client);
-
-		if (!cl->download.name.length())
-			continue;
-
-		static char buff[1024];
-		// Smaller buffer for slower clients
-		int chunk_size = (sizeof(buff) > (unsigned)cl->rate * 1000 / TICRATE) ? cl->rate * 1000 / TICRATE : sizeof(buff);
-
-		// maximum rate client can download at (in bytes per second)
-		int download_rate = (sv_waddownloadcap > cl->rate) ? cl->rate * 1000 : sv_waddownloadcap * 1000;
-
-		do
-		{
-			// read next bit of wad
-			unsigned int read;
-			unsigned int filelen = 0;
-			read = W_ReadChunk(cl->download.name.c_str(), cl->download.next_offset, chunk_size, buff, filelen);
-
-			if (!read)
-				break;
-
-			// [SL] 2011-08-09 - Always send the data in netbuf and reliablebuf prior
-			// to writing a wadchunk to netbuf to keep packet sizes below the MTU.
-			// This prevents packets from getting dropped due to size on some networks.
-			if (cl->netbuf.size() + cl->reliablebuf.size())
-				SV_SendPacket(*it);
-
-			if (!cl->download.next_offset)
-			{
-				MSG_WriteMarker(&cl->netbuf, svc_wadinfo);
-				MSG_WriteLong(&cl->netbuf, filelen);
-			}
-
-			MSG_WriteMarker(&cl->netbuf, svc_wadchunk);
-			MSG_WriteLong(&cl->netbuf, cl->download.next_offset);
-			MSG_WriteShort(&cl->netbuf, read);
-			MSG_WriteChunk(&cl->netbuf, buff, read);
-
-			// Make double-sure the wadchunk is sent in its own packet
-			if (cl->netbuf.size() + cl->reliablebuf.size())
-				SV_SendPacket(*it);
-
-			cl->download.next_offset += read;
-		} while (
-			(double)(cl->reliable_bps + cl->unreliable_bps) * TICRATE
-			/ (double)(gametic % TICRATE)	// bps already used
-			+ (double)chunk_size / TICRATE 	// bps this chunk will use
-			< (double)download_rate);
-	}
-}
-
-//
-// SV_WadDownloads
-//
-void SV_WadDownloads (void)
-{
-	if (sv_download_test)
-		SV_OptimizedDownloadTest();
-	else
-		SV_DownloadOriginal();
-}
 
 //
 //	SV_WinningTeam					[Toke - teams]
@@ -4834,8 +4679,6 @@ void SV_GameTics (void)
 
 	for (Players::iterator it = players.begin(); it != players.end(); ++it)
 		SV_ProcessPlayerCmd(*it);
-
-	SV_WadDownloads();
 }
 
 void SV_TouchSpecial(AActor *special, player_t *player)
